Unify the rotate left and right methods for add and erase
This deletes the rotate left and right for erase and instead makes
the method of rotate to receive a stop function instead of having
two methods whose only differences are the calling of a function.
Now you send it either isOverloaded or isUnderloaded depending on
you wanting to add or to erase.