Divide rotation for add and erase
Because the rotation to erase was wrong, it was necessary to remake
them and do something specific when you are going to add and when
you are going to delete. For that, a generic rotate left and right
was made and the rotate left and right for add and erase make use
of it with different conditions.
To accomplish the above mentioned, it was also needed to create a
method to get an iterator of a node.
Also, I thought I'd need a method to get an iterator to the left
or right sibling but in the end it wasn't needed. I'll erase them
if they prove to be useless afterwards.
Finally, I made some changes in main to be able to ask for a more
specific tree.