Change to mergeRoot method
Because the root doesn't exactly merge because it doesn't have an
ancestor to merge with, you can't really say the root merges but it's
children do. Because of this, now the method is mergeRootChildren an
receives the children to merge.
Because of this, the condition to call this function also changes so
searchSpaceErase also changed. Now instead of handling underloading
until the node is null (handled the root and now are a level above it),
it checks for the node to be the root, and if it is, then it stops.