Code mergeRootChildren and change call conditions
Now mergeRootChildren has code. Look at it for more details.
The conditions to call this method changed because the method also
changed its functionality. Now the method handleUnderload searches
for space for the node (because it should never be the root, the
only case it is invalidated), then it merges if it isn't a children
of the root or merges with the special method of childrens of the root.

Other change made was that now merge checks inside of it if a node is
leftmost rightmost or none and acts accordingly. This is also better
for the mergeRootChildren method because it calls the method and now
it doesn't have to check what type of merge should be called.