Fix merge and loop condition of getGreaterMinor
Made some changes to merge so it at least work with the current main
file.
With getGreaterMinor a condition was added so it doesn't loop
forever in the case where the received node doesn't have keys because
it's only one was erased.

Also erased an unnecessary variable in the areLeftSiblingAtMinumum
method.