Change erase and related methods and main file
Fixed a bug in erase where it sometimes deleted two times the same
information, causing a segmentation error. Some other methods also
changed in order to fix that.
In order to fix the above, quite a lot of changes had to be made.

Main file changed in an attempt to make it easier to test whether
the tree works or not.