Change trivial stuff
Fixed the find function to it correctly tells you whether there is
an element in the tree or not.
Changed checking if the list of children is empty for checking if
the node is a leaf in findPlace and findPlaceErase.
Added a comparison function in merge of children because apparently
merge needs to know how to compare the elements. Doesn't look like
it changed anything, though.