Add text files add.txt and erase.txt
Instead of having a unique default file for both adding and erasing,
an add file to add elements and an erase file to erase elements was
added.
To couple with this, all methods of the tree which are related to text
files also changed to have by default the filepath of add of they
add, or are generateFile or testAddAndDelete and erase if they are
related to erase.