Optimize Node related code and remove methods
-Node now is separated in declaration and implementation.
-Some virtual methods of node were made non virtual because they are
the same for all the child classes.
-The methods to get a key or child using an index were removed because
of the high cost of calling it multiple times. Use iterators instead.
-The variable rootFlag was removed. It was simply not needed. Now
isRoot is a virtual method and the child classes return true if are
the RootNode class or false if not.
-addItem now doesn't put a key and then sort because that's not efficient.
It now searches the place to put the key and inserts it there.
-addChild method was added. Think of it as an addItem for children.
-Makefile changed so it now compiles with the new implementation file
Node.cpp.