Stop using sort in BStarTree.cpp
Instead of using sort, it was replaced with addItem or addChild to
do add elements more efficiently. Other times the timing of inserting
an element was changed so it wasn't necessary to sort. In others it
was simply not needed.

Removed some unecessary conditionals and comments.

Improved the way to generate a list of non repetitive random numbers
in generateFile.