Calculate only once the limit of keys
In a lot of methods the minimum and maximum were recalculated when
only calculating once and using that value is enough. So now the
maximum and minimum number of keys for the root and normal node and
the number of keys to give to the children when splitting are calculated
only once and stored in variables in the BStarTree class where you can
see them by calling a method.
Because of the above, now it isn't necessary to store the order in the
nodes but it is to store the tree whose order you care about, so now
the nodes store a reference to a constant tree.