Remove repetitive lambdas
Instead now Node has 6 new methods: popFrontKeys, popBackKeys,
popFrontChild, popBackChild, putKeys, putChildren. The last two
are practically the same as the lambdas and the first four are a
shortcut to not have to get the front or back of a list and then
having to pop it. Those four methods do those two actions for you.